pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.2",
    "qrcode>=8.2",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]
